% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{quickAnalysis}
\alias{quickAnalysis}
\title{quickAnalysis Function}
\usage{
quickAnalysis(expres, groupingVar, min, sust, aFName = "Results",
  outputDir = ".", outputFType = NULL, pvalThreshold = 1,
  useAdjP = TRUE, plotSelected = TRUE, plot2pdf = FALSE)
}
\arguments{
\item{expres}{a matrix-like data object containing log-expression or numeric values, with rows corresponding to genes and columns to samples.}

\item{groupingVar}{factor which codes the grouping to be tested. There must be 1 or 2 groups.The length of the factor needs to correspond to the sample size.}

\item{min}{minuend}

\item{sust}{subtrahend}

\item{aFName}{the name to the output file.}

\item{outputDir}{the path to the output file.}

\item{outputFType}{character string with the abbreviation for extension: "xls" or "html. Default value is NULL.}

\item{pvalThreshold}{setting the threshold p-value. Default value is 1.}

\item{useAdjP}{a logical value indicating use adjust p-value. Default value is TRUE.}

\item{plotSelected}{A logical value indicating whether the output is a plot. Default value is TRUE.}

\item{plot2pdf}{A logical value indicating whether the plot is a pdf. Default value is FALSE.}
}
\value{
genes: selected genes

resT: multtest

topTable: limma
}
\description{
La función quickAnalysis realiza para una matriz de datos donde las filas corresponden a los genes y las columnas a las muestras un
análisis de expresión diferencial entre dos grupos concretos. Para el análisis usa dos métodos: un t.test para comparaciones multiples
(tabla resultante: reT) y un análisis limma básico (tabla resultante: topTable). Además, para aquellos genes diferencialmente expresados
realiza los boxplots para cada uno de los grupos.
}
\examples{
par(mfrow=c(2,2))
quickAnalysis(expres = exprs(sampleSet),
groupingVar = pData(sampleSet)$group,
min = "Case", sust = "Control", pvalThreshold = 0.25)
}
\author{
Alex Sánchez \email{alex.sanchez@vhir.org} , Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{differential}
\keyword{expression}
\keyword{limma}
\keyword{microarrays}
\keyword{multtest}
\keyword{plot}
\keyword{quickAnalysis}
